+++ apps/opentelemetry/src/otel_resource.erl
@@ -19,23 +19,33 @@
 -module(otel_resource).
 
 -export([create/1,
+         create/2,
          merge/2,
+         schema_url/1,
          attributes/1]).
 
--type key() :: unicode:latin1_chardata().
--type value() :: unicode:latin1_chardata() | integer() | float() | boolean().
--type resource() :: {otel_resource, otel_attributes:t()}.
--opaque t() :: resource().
+-include("otel_resource.hrl").
 
--export_type([t/0]).
+-type key() :: unicode:latin1_binary().
+%% values allowed in attributes of a resource are limited
+-type value() :: unicode:latin1_binary() | integer() | float() | boolean().
+-type schema_url() :: uri_string:uri_string().
 
 -define(MAX_LENGTH, 255).
 
-%% verifies each key and value and drops any that don't pass verification
+-record(resource, {schema_url :: schema_url() | undefined,
+                   attributes :: otel_attributes:t()}).
+-type t() :: #resource{}.
+
 -spec create(#{key() => value()} | [{key(), value()}]) -> t().
-create(Map) when is_map(Map) ->
-    create(maps:to_list(Map));
-create(List) when is_list(List) ->
+create(Attributes) ->
+    create(Attributes, undefined).
+
+%% verifies each key and value and drops any that don't pass verification
+-spec create(#{key() => value()} | [{key(), value()}], schema_url() | undefined) -> t().
+create(Map, SchemaUrl) when is_map(Map) ->
+    create(maps:to_list(Map), SchemaUrl);
+create(List, SchemaUrl) when is_list(List) ->
     List1 = lists:filtermap(fun({K, V}) ->
                                     %% TODO: log an info or debug message when dropping?
                                     case check_key(K) andalso check_value(V) of
@@ -45,19 +55,42 @@ create(List) when is_list(List) ->
                                             false
                                     end
                             end, lists:ukeysort(1, List)),
-    {otel_resource, otel_attributes:new(List1, 128, 255)}.
+    #resource{schema_url=SchemaUrl,
+              attributes=otel_attributes:new(List1, 128, 255)}.
+
+-spec schema_url(t()) -> schema_url() | undefined.
+schema_url(#resource{schema_url=Schema}) ->
+    Schema.
 
 -spec attributes(t()) -> otel_attributes:t().
-attributes({otel_resource, Attributes}) ->
+attributes(#resource{attributes=Attributes}) ->
     Attributes.
 
-%% In case of collision the updating, first argument, resource takes precedence.
+%% in case of collision the updating, first argument, resource takes precedence.
 -spec merge(t(), t()) -> t().
-merge({otel_resource, NewAttributes}, {otel_resource, CurrentAttributes}) ->
+merge(#resource{schema_url=NewSchemaUrl,
+                attributes=NewAttributes}, Current=#resource{schema_url=CurrentSchemaUrl,
+                                                             attributes=CurrentAttributes}) ->
+    SchameUrl = merge_schema_url(NewSchemaUrl, CurrentSchemaUrl),
     NewMap = otel_attributes:map(NewAttributes),
-    {otel_resource, otel_attributes:set(NewMap, CurrentAttributes)}.
+    Current#resource{schema_url=SchameUrl,
+                     attributes=otel_attributes:set(NewMap, CurrentAttributes)}.
 
 %%
+
+%% when merging resources the schemas are checked to verify they match.
+%% if they do match then the schema is set to `undefined' and the attributes are kept.
+merge_schema_url(SchemaUrl, undefined) ->
+    SchemaUrl;
+merge_schema_url(undefined, SchemaUrl) ->
+    SchemaUrl;
+merge_schema_url(NewSchemaUrl, CurrentSchemaUrl) ->
+    merge_schema_url_(uri_string:normalize(NewSchemaUrl), uri_string:normalize(CurrentSchemaUrl)).
+
+merge_schema_url_(SchemaUrl, SchemaUrl) ->
+    SchemaUrl;
+merge_schema_url_(_, _) ->
+    undefined.
 
 %% all resource strings, key or value, must be latin1 with length less than 255
 check_string(S) ->