+++ apps/opentelemetry_api/lib/open_telemetry.ex
@@ -116,7 +116,7 @@ defmodule OpenTelemetry do
   @type status() :: :opentelemetry.status()
 
   defdelegate get_tracer(name), to: :opentelemetry
-  defdelegate register_tracer(name, vsn), to: :opentelemetry
+  defdelegate get_tracer(name, vsn, schema_url), to: :opentelemetry
   defdelegate set_default_tracer(t), to: :opentelemetry
 
   # Helpers to build OpenTelemetry structured types