+++ apps/opentelemetry_api/src/opentelemetry.erl
@@ -29,10 +29,10 @@
 -module(opentelemetry).
 
 -export([set_default_tracer/1,
-         register_tracer/2,
-         register_applications/1,
+         create_application_tracers/1,
          get_tracer/0,
          get_tracer/1,
+         get_tracer/3,
          set_tracer/2,
          get_application/1,
          get_application_tracer/1,
@@ -41,7 +41,7 @@
          get_text_map_extractor/0,
          set_text_map_injector/1,
          get_text_map_injector/0,
-         instrumentation_library/2,
+         instrumentation_library/3,
          timestamp/0,
          timestamp_to_nano/1,
          convert_timestamp/2,
@@ -148,48 +148,84 @@ set_default_tracer(Tracer) ->
 
 -spec set_tracer(atom(), tracer()) -> boolean().
 set_tracer(Name, Tracer) ->
-    verify_and_set_term(Tracer, ?TRACER_KEY(Name), otel_tracer).
+    set_tracer(Name, <<>>, undefined, Tracer).
 
--spec register_tracer(atom(), string() | binary()) -> boolean().
-register_tracer(Name, Vsn) when is_atom(Name) and is_binary(Vsn) ->
-    otel_tracer_provider:register_tracer(Name, Vsn);
-register_tracer(Name, Vsn) when is_atom(Name) and is_list(Vsn) ->
-    otel_tracer_provider:register_tracer(Name, list_to_binary(Vsn)).
+-spec set_tracer(Name, Vsn, SchemaUrl, Tracer) -> boolean() when
+      Name :: atom(),
+      Vsn :: unicode:chardata() | undefined,
+      SchemaUrl :: uri_string:uri_string() | undefined,
+      Tracer:: opentelemetry:tracer().
+set_tracer(Name, Vsn, SchemaUrl, Tracer) ->
+    verify_and_set_term(Tracer, ?TRACER_KEY({Name, Vsn, SchemaUrl}), otel_tracer).
 
--spec register_applications([atom()]) -> ok.
-register_applications(Applications) ->
-    TracerMap = lists:foldl(fun(Application, Acc) ->
-                                    register_application_tracer(Application),
-                                    maps:merge(Acc, module_to_application(Application))
+-spec create_application_tracers([{Application, Description, Vsn}]) -> ok when
+      Application :: atom(),
+      Description :: string(),
+      Vsn :: string().
+create_application_tracers(Applications) ->
+    TracerMap = lists:foldl(fun({Name, _Description, Version}, Acc) ->
+                                    Vsn = vsn_to_binary(Version),
+                                    SchemaUrl = application:get_env(Name, otel_schema_url, undefined),
+                                    _ = get_tracer(Name, Version, SchemaUrl),
+                                    maps:merge(Acc, module_to_application(Name, Vsn, SchemaUrl))
                             end, #{}, Applications),
     persistent_term:put(?MODULE_TO_APPLICATION_KEY, TracerMap).
 
-register_application_tracer({Name, _Description, Version}) ->
-    register_tracer(Name, Version).
-
-%% creates a map of modules to application name
-module_to_application({Name, _Description, _Version}) ->
+%% creates a map of modules to application name, version and schema_url tuple
+module_to_application(Name, Version, SchemaUrl) ->
     {ok, Modules} = application:get_key(Name, modules),
     lists:foldl(fun(M, Acc) ->
-                        Acc#{M => Name}
+                        Acc#{M => {Name, Version, SchemaUrl}}
                 end, #{}, Modules).
 
 -spec get_tracer() -> tracer().
 get_tracer() ->
     persistent_term:get(?DEFAULT_TRACER_KEY, {otel_tracer_noop, []}).
 
--spec get_tracer(atom()) -> tracer().
+-spec get_tracer(Name) -> Tracer when
+      Name :: atom() | {Name, Vsn, SchemaUrl},
+      Vsn :: unicode:chardata() | undefined,
+      SchemaUrl :: uri_string:uri_string() | undefined,
+      Tracer:: opentelemetry:tracer().
+get_tracer('$__default_tracer') ->
+    get_tracer();
+get_tracer({Name, Vsn, SchemaUrl}) ->
+    get_tracer(Name, Vsn, SchemaUrl);
 get_tracer(Name) ->
-    persistent_term:get(?TRACER_KEY(Name), get_tracer()).
+    get_tracer(Name, undefined, undefined).
+
+-spec get_tracer(Name, Vsn, SchemaUrl) -> Tracer when
+      Name :: atom(),
+      Vsn :: unicode:chardata() | undefined,
+      SchemaUrl :: uri_string:uri_string() | undefined,
+      Tracer:: opentelemetry:tracer().
+get_tracer(Name, Vsn, SchemaUrl) ->
+    %% check cache and then use provider to get the tracer if it isn't cached yet
+    case persistent_term:get(?TRACER_KEY({Name, Vsn, SchemaUrl}), undefined) of
+        undefined ->
+            VsnBin = vsn_to_binary(Vsn),
+            Tracer = otel_tracer_provider:get_tracer(Name, VsnBin, SchemaUrl),
+
+            %% cache the tracer
+            _ = set_tracer(Name, Vsn, SchemaUrl, Tracer),
+
+            Tracer;
+        Tracer ->
+            Tracer
+    end.
 
 -spec get_application_tracer(module()) -> tracer().
 get_application_tracer(ModuleName) ->
     get_tracer(get_application(ModuleName)).
 
-%% looks up the name of the OTP Application a module is in. This name is used to
-%% look up a Tracer to use so if none is found for the ModuleName the key used for
-%% the default tracer
--spec get_application(module()) -> atom().
+%% looks up the name, version and schema_url used to create a Trace for the OTP
+%% Application a module is in. This name is used to look up a Tracer to use so
+%% if none is found for the ModuleName the key used for the default tracer.
+-spec get_application(module()) -> ApplicationTuple when
+      ApplicationTuple :: {Name, Vsn, SchemaUrl} | atom(),
+      Name :: atom(),
+      Vsn :: unicode:unicode_binary() | undefined,
+      SchemaUrl :: uri_string:uri_string() | undefined.
 get_application(ModuleName) ->
     Map = persistent_term:get(?MODULE_TO_APPLICATION_KEY, #{}),
     maps:get(ModuleName, Map, '$__default_tracer').
@@ -364,7 +400,7 @@ verify_and_set_term(Module, TermKey, Behaviour) ->
         false ->
             ?LOG_WARNING("Module ~p does not exist. "
                          "A noop ~p will be used until a module is configured.",
-                         [Module, Behaviour, Behaviour]),
+                         [Module, Behaviour]),
             false
     end.
 
@@ -390,18 +426,28 @@ link_or_false(TraceId, SpanId, Attributes, TraceState) ->
             false
     end.
 
-instrumentation_library(Name, Vsn) ->
-    #instrumentation_library{name=name_to_binary(Name),
-                             version=vsn_to_binary(Vsn)}.
+instrumentation_library(Name, Vsn, SchemaUrl) ->
+    case name_to_binary(Name) of
+        undefined ->
+            undefined;
+        BinaryName ->
+            #instrumentation_library{name=BinaryName,
+                                     version=vsn_to_binary(Vsn),
+                                     schema_url=schema_url_to_binary(SchemaUrl)}
+    end.
+
+%% schema_url is option, so set to undefined if its not a string
+schema_url_to_binary(SchemaUrl) when is_binary(SchemaUrl) ; is_list(SchemaUrl) ->
+    unicode:characters_to_binary(SchemaUrl);
+schema_url_to_binary(_) ->
+    undefined.
 
 %% Vsn can't be an atom or anything but a list or binary
-%% so return `undefined' if it isn't a list or binary.
-vsn_to_binary(Vsn) when is_list(Vsn) ->
-    list_to_binary(Vsn);
-vsn_to_binary(Vsn) when is_binary(Vsn) ->
-    Vsn;
+%% so return empty binary string if it isn't a list or binary.
+vsn_to_binary(Vsn) when is_binary(Vsn) ; is_list(Vsn) ->
+    unicode:characters_to_binary(Vsn);
 vsn_to_binary(_) ->
-    undefined.
+    <<>>.
 
 %% name can be atom, list or binary. But atom `undefined'
 %% must stay as `undefined' atom.