+++ apps/opentelemetry_experimental/src/otel_meter_server.erl
@@ -45,7 +45,7 @@ register_meter(Name, Vsn, #state{meter=Meter,
         true ->
             opentelemetry_experimental:set_meter(Name, {otel_meter_noop, []});
         false ->
-            InstrumentationLibrary = opentelemetry:instrumentation_library(Name, Vsn),
+            InstrumentationLibrary = opentelemetry:instrumentation_library(Name, Vsn, undefined),
             opentelemetry_experimental:set_meter(Name, {Meter#meter.module,
                                                         Meter#meter{instrumentation_library=InstrumentationLibrary}})
     end.